+++ producer.cpp
@@ -10,31 +10,42 @@
 #include "consumer.hpp"
 using namespace std;
 
-mutex mtx;
+// mutex zoomReport, studentReport, stuLock;
 
-struct student {
-    string email;
-    string startDate;
-    string startTime;
-    string endDate;
-    string endTime;
-};
+// struct student {
+//     string email;
+//     string startDate;
+//     string startTime;
+//     string endDate;
+//     string endTime;
+// };
 vector<student> stu;
 void parseText(string line, vector<student>& stu);
 
 int main(int argc, char *argv[]) {
   if (argc < 4) { cout << "Provide enough command line arguments"; }
-  consumer threaded(stoi(argv[3]), argv[2]);
+  consumer threaded(stoi(argv[3]), argv[2], stu);
   ifstream zoomreport;
   ofstream studentreport;
   string line;
-
+  zoomReport.lock();
   zoomreport.open("Zoomreport.txt");
+  studentReport.lock();
   studentreport.open("studentreport.txt");
-
+  stuLock.lock();
   while(getline(zoomreport, line)) {
       parseText(line, stu);
   }
+  stuLock.unlock();
+  zoomreport.close();
+  zoomReport.unlock();
+  stuLock.lock();
+  for(size_t i = 0; i < stu.size(); i++)
+    studentreport << stu[i].email << stu[i].startDate << stu[i].startTime << stu[i].endDate << stu[i].endTime << '\n';
+  stuLock.unlock();
+  studentreport.close();
+  studentReport.unlock();
+  threaded.execute();
   return 0;
 }
 void parseText(string line, vector<student>& stu) {
